+++ src/main/resources/application.properties
@@ -1,6 +1,6 @@
 server.servlet.session.tracking-modes=cookie
 
-spring.datasource.url=jdbc:h2:~/java-ims;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.url=jdbc:h2:mem:testdb
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
@@ -12,4 +12,4 @@ handlebars.suffix=.html
 handlebars.cache=false
 handlebars.expose-session-attributes=true
 
-file.upload.local=/Users/choi/dev/java-ims-file/
+file.upload.local=/home/ubuntu/attachments/
